Move `parse` to `pkg/parse`
